Move message panel around, set height
