+++ src/client/routes/game/GameView.tsx
@@ -3,14 +3,10 @@ import { withRouter, RouteComponentProps } from "react-router-dom"
 import { withStyles, WithStyles, createStyles, Theme } from "@material-ui/core/styles"
 import Board from "./Board"
 import { observer, inject } from "mobx-react"
-import { Button, Dialog, DialogTitle, DialogContent, DialogActions, Paper, Typography } from "@material-ui/core"
 import {IAppStore} from "../../store/AppStore"
 import GameStore from "../../store/GameStore"
-import {ApiPlayerDetails} from "../../../common/types"
-import SidePanel from "./SidePanel"
 import EndScreen from "./EndScreen"
 import MessagePanel from "./MessagePanel"
-import Players from "./Players"
 import PlayerBadge from "./PlayerBadge"
 import * as Majavashakki from "../../../common/GamePieces"
