+++ src/client/routes/game/MessagePanel/Message.tsx
@@ -2,8 +2,8 @@ import * as React from "react"
 import { withStyles, WithStyles } from "@material-ui/core/styles"
 import { observer, inject } from "mobx-react"
 import { Typography } from "@material-ui/core"
-import { IAppStore } from "../../store/AppStore";
-import * as Majavashakki from "../../../common/GamePieces"
+import { IAppStore } from "../../../store/AppStore";
+import * as Majavashakki from "../../../../common/GamePieces"
 
 class Message extends React.Component<IMessageProps, any> {
     public render() {
