+++ src/client/routes/game/MessagePanel/MessagePanel.tsx
@@ -5,7 +5,7 @@ import { IAppStore } from "client/store/AppStore"
 import { observer, inject } from "mobx-react"
 import GameStore from "client/store/GameStore"
 import Message from "./Message"
-import * as Majavashakki from "../../../common/GamePieces"
+import * as Majavashakki from "../../../../common/GamePieces"
 
 @inject((stores: IAppStore) => ({game: stores.app.game}))
 @observer
@@ -13,7 +13,7 @@ class MessagePanel extends React.Component<IMessagePanelProps, any> {
     public render() {
         console.log(this.props.game)
         return (
-            <Paper>
+            <Paper className={this.props.classes.container}>
                 {this.props.game.boardStore.moveHistory.map((move, index) => {
                     console.log("test")
                     const color = index % 2 === 0 ? Majavashakki.PieceColor.Black : Majavashakki.PieceColor.White
@@ -30,6 +30,10 @@ interface IMessagePanelProps extends WithStyles<typeof styles> {
     game?: GameStore,
 }
 
-const styles = () => createStyles({ })
+const styles = () => createStyles({
+  container: {
+    height: "20vmin",
+  },
+})
 
 export default withStyles(styles)(MessagePanel)
