+++ src/client/routes/game/MessagePanel/index.ts
@@ -1,0 +1,2 @@
+import MessagePanel from "./MessagePanel"
+export default MessagePanel
